Refactor network manager
