+++ paykit/src/test/java/com/squareup/cash/paykit/CashAppPayKitAuthorizeTests.kt
@@ -29,21 +29,21 @@ class CashAppPayKitAuthorizeTests {
 
   @Test(expected = PayKitIntegrationException::class)
   fun `should throw if calling authorize before createCustomer`() {
-    val payKit = CashAppPayKitImpl(FakeData.CLIENT_ID, useSandboxEnvironment = true)
+    val payKit = createPayKit()
     payKit.registerForStateUpdates(mockk())
     payKit.authorizeCustomerRequest(context)
   }
 
   @Test(expected = PayKitIntegrationException::class)
   fun `should throw on authorizeCustomerRequest if has NOT registered for state updates`() {
-    val payKit = CashAppPayKitImpl(FakeData.CLIENT_ID, useSandboxEnvironment = true)
+    val payKit = createPayKit()
     val customerResponseData = mockk<CustomerResponseData>(relaxed = true)
     payKit.authorizeCustomerRequest(context, customerResponseData)
   }
 
   @Test(expected = IllegalArgumentException::class)
   fun `should throw if missing mobileUrl from customer data`() {
-    val payKit = CashAppPayKitImpl(FakeData.CLIENT_ID, useSandboxEnvironment = true)
+    val payKit = createPayKit()
     val customerResponseData = mockk<CustomerResponseData>(relaxed = true)
     payKit.registerForStateUpdates(mockk())
 
@@ -52,7 +52,7 @@ class CashAppPayKitAuthorizeTests {
 
   @Test(expected = IllegalArgumentException::class)
   fun `should throw if unable to parse mobile url in customer data`() {
-    val payKit = CashAppPayKitImpl(FakeData.CLIENT_ID, useSandboxEnvironment = true)
+    val payKit = createPayKit()
     val customerResponseData = mockk<CustomerResponseData>(relaxed = true) {
       every { authFlowTriggers } returns null
     }
@@ -63,7 +63,7 @@ class CashAppPayKitAuthorizeTests {
 
   @Test(expected = RuntimeException::class)
   fun `should throw on if unable to start mobileUrl activity`() {
-    val payKit = CashAppPayKitImpl(FakeData.CLIENT_ID, useSandboxEnvironment = true)
+    val payKit = createPayKit()
     val customerResponseData = mockk<CustomerResponseData>(relaxed = true) {
       every { authFlowTriggers } returns mockk {
         every { mobileUrl } returns "http://url"
@@ -73,4 +73,11 @@ class CashAppPayKitAuthorizeTests {
 
     payKit.authorizeCustomerRequest(context, customerResponseData)
   }
+
+  private fun createPayKit() =
+    CashAppPayKitImpl(
+      clientId = FakeData.CLIENT_ID,
+      networkManager = mockk(),
+      useSandboxEnvironment = true,
+    )
 }
