+++ paykit/src/test/java/com/squareup/cash/paykit/NetworkErrorTests.kt
@@ -4,6 +4,7 @@ import com.squareup.cash.paykit.PayKitState.PayKitException
 import com.squareup.cash.paykit.exceptions.PayKitApiNetworkException
 import com.squareup.cash.paykit.exceptions.PayKitConnectivityNetworkException
 import com.squareup.cash.paykit.impl.CashAppPayKitImpl
+import com.squareup.cash.paykit.impl.NetworkManagerImpl
 import com.squareup.moshi.JsonDataException
 import okhttp3.mockwebserver.MockResponse
 import okhttp3.mockwebserver.MockWebServer
@@ -18,10 +19,6 @@ class NetworkErrorTests {
 
   @Test
   fun `HTTP code without payload should be wrapped with correct SDK defined exception`() {
-    val payKit = CashAppPayKitImpl(FakeData.CLIENT_ID, useSandboxEnvironment = true)
-    val mockListener = MockListener()
-    payKit.registerForStateUpdates(mockListener)
-
     // Setup server & mock responses.
     val server = MockWebServer()
     server.enqueue(MockResponse().setResponseCode(503))
@@ -30,7 +27,12 @@ class NetworkErrorTests {
     server.start()
 
     val baseUrl = server.url("")
-    NetworkManager.baseUrl = baseUrl.toString()
+
+    val networkManager = NetworkManagerImpl(baseUrl = baseUrl.toString())
+    val payKit = createPayKit(networkManager)
+    val mockListener = MockListener()
+    payKit.registerForStateUpdates(mockListener)
+
     payKit.createCustomerRequest(FakeData.oneTimePayment)
 
     // Verify that all the appropriate exception wrapping has occurred for a 503 error.
@@ -47,10 +49,6 @@ class NetworkErrorTests {
 
   @Test
   fun `HTTP error code with payload should be wrapped with correct SDK defined exception and contains API deserialized data`() {
-    val payKit = CashAppPayKitImpl(FakeData.CLIENT_ID, useSandboxEnvironment = true)
-    val mockListener = MockListener()
-    payKit.registerForStateUpdates(mockListener)
-
     // Setup server & mock responses.
     val server = MockWebServer()
     server.enqueue(
@@ -72,7 +70,11 @@ class NetworkErrorTests {
     server.start()
 
     val baseUrl = server.url("")
-    NetworkManager.baseUrl = baseUrl.toString()
+    val networkManager = NetworkManagerImpl(baseUrl = baseUrl.toString())
+    val payKit = createPayKit(networkManager)
+    val mockListener = MockListener()
+    payKit.registerForStateUpdates(mockListener)
+
     payKit.createCustomerRequest(FakeData.oneTimePayment)
 
     // Verify that all the appropriate exception wrapping has occurred for a 400 error.
@@ -91,9 +93,6 @@ class NetworkErrorTests {
 
   @Test
   fun `network request timeout should be wrapped with correct SDK defined exception`() {
-    val payKit = CashAppPayKitImpl(FakeData.CLIENT_ID, useSandboxEnvironment = true)
-    val mockListener = MockListener()
-    payKit.registerForStateUpdates(mockListener)
     // Setup server & mock responses.
     val server = MockWebServer()
     server.enqueue(
@@ -103,11 +102,14 @@ class NetworkErrorTests {
     // Start the server.
     server.start()
     val baseUrl = server.url("")
-    NetworkManager.baseUrl = baseUrl.toString()
-    NetworkManager.DEFAULT_NETWORK_TIMEOUT_MILLISECONDS = 1
+
+    val networkManager =
+      NetworkManagerImpl(baseUrl = baseUrl.toString(), networkTimeoutMilliseconds = 1)
+    val payKit = createPayKit(networkManager)
+    val mockListener = MockListener()
+    payKit.registerForStateUpdates(mockListener)
 
     payKit.createCustomerRequest(FakeData.oneTimePayment)
-
     // Verify that a timeout error was captured and relayed to the SDK listener.
     assertTrue(
       "Expected SocketTimeoutException",
@@ -117,10 +119,6 @@ class NetworkErrorTests {
 
   @Test
   fun `unsupported json payload should be wrapped in corresponding SDK exception`() {
-    val payKit = CashAppPayKitImpl(FakeData.CLIENT_ID, useSandboxEnvironment = true)
-    val mockListener = MockListener()
-    payKit.registerForStateUpdates(mockListener)
-
     // Setup server & mock responses.
     val server = MockWebServer()
 
@@ -165,7 +163,11 @@ class NetworkErrorTests {
     server.start()
 
     val baseUrl = server.url("")
-    NetworkManager.baseUrl = baseUrl.toString()
+    val networkManager = NetworkManagerImpl(baseUrl = baseUrl.toString())
+    val payKit = createPayKit(networkManager)
+    val mockListener = MockListener()
+    payKit.registerForStateUpdates(mockListener)
+
     payKit.createCustomerRequest(FakeData.oneTimePayment)
 
     // Verify that we got the appropriate JSON deserialization error.
@@ -186,4 +188,11 @@ class NetworkErrorTests {
       state = newState
     }
   }
+
+  private fun createPayKit(networkManager: NetworkManager) =
+    CashAppPayKitImpl(
+      clientId = FakeData.CLIENT_ID,
+      networkManager = networkManager,
+      useSandboxEnvironment = true,
+    )
 }
